test(architecture): move backend auth infrastructure tests
