+++ test/unit/infrastructure/trusted-origin.test.ts
@@ -1,4 +1,4 @@
-﻿import test from "node:test";
+import test from "node:test";
 import assert from "node:assert/strict";
 
 process.env.SUPABASE_URL ??= "https://example.supabase.co";
@@ -7,8 +7,8 @@ process.env.SUPABASE_SERVICE_ROLE_KEY ??= "test-service-role-key";
 process.env.DATABASE_URL ??= "postgresql://postgres:postgres@127.0.0.1:5432/postgres";
 process.env.SUPABASE_DB_URL ??= process.env.DATABASE_URL;
 
-const { ENV } = await import("../server/lib/env.ts");
-const { requireTrustedOrigin } = await import("../server/middlewares/trusted-origin.ts");
+const { ENV } = await import("../../../server/lib/env.ts");
+const { requireTrustedOrigin } = await import("../../../server/middlewares/trusted-origin.ts");
 
 function createMockResponse() {
   return {
